Extract Image upload Views and Models to own files
Now that the views and models don't need to share any local variables or
manipulate the DOM when being defined, we can use the same conventions
as our other Backbone files.
