+++ backend/app/assets/javascripts/spree/backend/models/index.js
@@ -1,2 +1,3 @@
+//= require spree/backend/models/image_upload
 //= require spree/backend/models/line_item
 //= require spree/backend/models/order
