+++ backend/app/assets/javascripts/spree/backend/namespaces.js
@@ -2,6 +2,7 @@ _.extend(window.Spree, {
   Models: {},
   Collections: {},
   Views: {
+    Images: {},
     Order: {},
     Cart: {},
     Zones: {}
