+++ backend/app/assets/javascripts/spree/backend/views/index.js
@@ -1,6 +1,8 @@
 //= require 'spree/backend/views/cart/add_line_item_button'
 //= require 'spree/backend/views/cart/line_item_row'
 //= require 'spree/backend/views/cart/line_item_table'
+//= require 'spree/backend/views/images/upload_zone'
+//= require 'spree/backend/views/images/upload_progress'
 //= require 'spree/backend/views/order/address'
 //= require 'spree/backend/views/order/details_adjustments'
 //= require 'spree/backend/views/order/details_total'
